pyproject.toml
[project]
name = "triage-cot-labeling"
version = "0.1.0"
description = "Wiki-guided JSON annotation pipeline for triage images."
requires-python = ">=3.10"
dependencies = [
  "google-genai",
  "pyyaml",
]

[tool.uv]
package = false
